Move tag related function in it's own common tag handler
