+++ backend/src/main/java/ai/verta/modeldb/experiment/ExperimentServiceImpl.java
@@ -15,6 +15,7 @@ import ai.verta.modeldb.common.event.FutureEventDAO;
 import ai.verta.modeldb.common.exceptions.AlreadyExistsException;
 import ai.verta.modeldb.common.exceptions.ModelDBException;
 import ai.verta.modeldb.common.exceptions.NotFoundException;
+import ai.verta.modeldb.common.handlers.TagsHandlerBase;
 import ai.verta.modeldb.exceptions.InvalidArgumentException;
 import ai.verta.modeldb.exceptions.PermissionDeniedException;
 import ai.verta.modeldb.metadata.MetadataServiceImpl;
@@ -135,7 +136,7 @@ public class ExperimentServiceImpl extends ExperimentServiceImplBase {
             .setName(ModelDBUtils.checkEntityNameLength(request.getName()))
             .setDescription(request.getDescription())
             .addAllAttributes(request.getAttributesList())
-            .addAllTags(CommonUtils.checkEntityTagsLength(request.getTagsList()))
+            .addAllTags(TagsHandlerBase.checkEntityTagsLength(request.getTagsList()))
             .addAllArtifacts(request.getArtifactsList())
             .setVersionNumber(1L);
 
@@ -512,7 +513,7 @@ public class ExperimentServiceImpl extends ExperimentServiceImplBase {
 
       var updatedExperiment =
           experimentDAO.addExperimentTags(
-              request.getId(), CommonUtils.checkEntityTagsLength(request.getTagsList()));
+              request.getId(), TagsHandlerBase.checkEntityTagsLength(request.getTagsList()));
       var response =
           AddExperimentTags.Response.newBuilder().setExperiment(updatedExperiment).build();
 
@@ -566,7 +567,7 @@ public class ExperimentServiceImpl extends ExperimentServiceImplBase {
       var updatedExperiment =
           experimentDAO.addExperimentTags(
               request.getId(),
-              CommonUtils.checkEntityTagsLength(Collections.singletonList(request.getTag())));
+              TagsHandlerBase.checkEntityTagsLength(Collections.singletonList(request.getTag())));
       var response =
           AddExperimentTag.Response.newBuilder().setExperiment(updatedExperiment).build();
