+++ backend/src/main/java/ai/verta/modeldb/versioning/CommitDAORdbImpl.java
@@ -8,10 +8,10 @@ import ai.verta.modeldb.DatasetVersion;
 import ai.verta.modeldb.ModelDBConstants;
 import ai.verta.modeldb.PathLocationTypeEnum.PathLocationType;
 import ai.verta.modeldb.authservice.MDBRoleService;
-import ai.verta.modeldb.common.CommonUtils;
 import ai.verta.modeldb.common.authservice.AuthService;
 import ai.verta.modeldb.common.collaborator.CollaboratorUser;
 import ai.verta.modeldb.common.exceptions.ModelDBException;
+import ai.verta.modeldb.common.handlers.TagsHandlerBase;
 import ai.verta.modeldb.cron_jobs.DeleteEntitiesCron;
 import ai.verta.modeldb.dto.CommitPaginationDTO;
 import ai.verta.modeldb.entities.AttributeEntity;
@@ -872,10 +872,11 @@ public class CommitDAORdbImpl implements CommitDAO {
               .setStringId(compositeId)
               .build();
       if (addLabels) {
-        metadataDAO.addLabels(identificationType, CommonUtils.checkEntityTagsLength(labelsList));
+        metadataDAO.addLabels(
+            identificationType, TagsHandlerBase.checkEntityTagsLength(labelsList));
       } else {
         metadataDAO.deleteLabels(
-            identificationType, CommonUtils.checkEntityTagsLength(labelsList), deleteAll);
+            identificationType, TagsHandlerBase.checkEntityTagsLength(labelsList), deleteAll);
       }
       var commitEntity = getCommitEntity(session, commitHash, (session1 -> repositoryEntity));
       commitEntity.setDate_updated(new Date().getTime());
